Add header to top-list
