+++ examples/top-list/package.json
@@ -6,6 +6,7 @@
     "project-name-generator": "^2.1.4",
     "react": "^16.2.0",
     "react-addons-css-transition-group": "^15.6.2",
+    "react-bootstrap": "^0.32.1",
     "react-dom": "^16.2.0",
     "react-helmet": "^5.2.0",
     "react-redux": "^5.0.6",
